requirements.txt
asgiref==3.2.3
automium==0.2.6
automium-web==0.1.1
Click==7.0
DateTime==4.3
Django==3.0.3
djangorestframework==3.11.0
Flask==1.1.1
itsdangerous==1.1.0
Jinja2==2.11.1
legacy==0.1.6
MarkupSafe==1.1.1
pytz==2019.3
sqlparse==0.3.0
Werkzeug==1.0.0
zope.interface==4.7.1
